refactor: re-name (`beveragesMap` to `myStock`)
